+++ Ruoyi/ruoyi-DS/src/main/java/com/ruoyi/DS/service/IOrderSkuService.java
@@ -17,7 +17,7 @@ public interface IOrderSkuService
      * @param id 单品项ID
      * @return 单品项
      */
-    public OrderSku selectOrderSkuById(Long id);
+    public OrderSku selectOrderSkuById(Integer id);
 
     /**
      * 查询单品项列表
@@ -57,5 +57,5 @@ public interface IOrderSkuService
      * @param id 单品项ID
      * @return 结果
      */
-    public int deleteOrderSkuById(Long id);
+    public int deleteOrderSkuById(Integer id);
 }
