+++ Ruoyi/ruoyi-DS/src/main/java/com/ruoyi/DS/service/ISalesOrderService.java
@@ -57,7 +57,7 @@ public interface ISalesOrderService
      * @param id 销售订单ID
      * @return 结果
      */
-    public int deleteSalesOrderById(Long id);
+    public int deleteSalesOrderById(Integer id);
