+++ Ruoyi/ruoyi-admin/src/main/java/com/ruoyi/web/controller/DS/OrderSkuController.java
@@ -1,15 +1,13 @@
 package com.ruoyi.web.controller.DS;
 
 import java.util.List;
+
+import com.ruoyi.DS.domain.Skuproduct;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 import com.ruoyi.common.annotation.Log;
 import com.ruoyi.common.enums.BusinessType;
 import com.ruoyi.DS.domain.OrderSku;
@@ -27,8 +25,7 @@ import com.ruoyi.common.core.page.TableDataInfo;
  */
 @Controller
 @RequestMapping("/DS/order_sku")
-public class OrderSkuController extends BaseController
-{
+public class OrderSkuController extends BaseController {
     private String prefix = "DS/order_sku";
 
     @Autowired
@@ -36,8 +33,7 @@ public class OrderSkuController extends BaseController
 
     @RequiresPermissions("DS:order_sku:view")
     @GetMapping()
-    public String order_sku()
-    {
+    public String order_sku() {
         return prefix + "/order_sku";
     }
 
@@ -47,8 +43,7 @@ public class OrderSkuController extends BaseController
     @RequiresPermissions("DS:order_sku:list")
     @PostMapping("/list")
     @ResponseBody
-    public TableDataInfo list(OrderSku orderSku)
-    {
+    public TableDataInfo list(OrderSku orderSku) {
         startPage();
         List<OrderSku> list = orderSkuService.selectOrderSkuList(orderSku);
         return getDataTable(list);
@@ -61,8 +56,7 @@ public class OrderSkuController extends BaseController
     @Log(title = "单品项", businessType = BusinessType.EXPORT)
     @PostMapping("/export")
     @ResponseBody
-    public AjaxResult export(OrderSku orderSku)
-    {
+    public AjaxResult export(OrderSku orderSku) {
         List<OrderSku> list = orderSkuService.selectOrderSkuList(orderSku);
         ExcelUtil<OrderSku> util = new ExcelUtil<OrderSku>(OrderSku.class);
         return util.exportExcel(list, "order_sku");
@@ -72,8 +66,7 @@ public class OrderSkuController extends BaseController
      * 新增单品项
      */
     @GetMapping("/add")
-    public String add()
-    {
+    public String add() {
         return prefix + "/add";
     }
 
@@ -84,8 +77,7 @@ public class OrderSkuController extends BaseController
     @Log(title = "单品项", businessType = BusinessType.INSERT)
     @PostMapping("/add")
     @ResponseBody
-    public AjaxResult addSave(OrderSku orderSku)
-    {
+    public AjaxResult addSave(OrderSku orderSku) {
         return toAjax(orderSkuService.insertOrderSku(orderSku));
     }
 
@@ -93,8 +85,7 @@ public class OrderSkuController extends BaseController
      * 修改单品项
      */
     @GetMapping("/edit/{id}")
-    public String edit(@PathVariable("id") Long id, ModelMap mmap)
-    {
+    public String edit(@PathVariable("id") Integer id, ModelMap mmap) {
         OrderSku orderSku = orderSkuService.selectOrderSkuById(id);
         mmap.put("orderSku", orderSku);
         return prefix + "/edit";
@@ -107,8 +98,7 @@ public class OrderSkuController extends BaseController
     @Log(title = "单品项", businessType = BusinessType.UPDATE)
     @PostMapping("/edit")
     @ResponseBody
-    public AjaxResult editSave(OrderSku orderSku)
-    {
+    public AjaxResult editSave(OrderSku orderSku) {
         return toAjax(orderSkuService.updateOrderSku(orderSku));
     }
 
@@ -117,10 +107,13 @@ public class OrderSkuController extends BaseController
      */
     @RequiresPermissions("DS:order_sku:remove")
     @Log(title = "单品项", businessType = BusinessType.DELETE)
-    @PostMapping( "/remove")
+    @PostMapping("/remove")
     @ResponseBody
-    public AjaxResult remove(String ids)
-    {
+    public AjaxResult remove(String ids) {
         return toAjax(orderSkuService.deleteOrderSkuByIds(ids));
     }
+
+
+
+
 }
